backend/requirements.txt
google-genai==1.52.0
python-dotenv==1.2.1
supabase==2.24.0
fastapi==0.122.0
uvicorn==0.38.0
